* develop:
  Fixed issue with staff.csv
  Replaced Personnel with Officer for Role
  Show only corresponding department for Agency FOIL Officer
  Check that record.filename is not None
  Show Department Filter for Agency FOIL Officer

+++ public_records_portal/prr.py
@@ -289,7 +289,7 @@ def update_resource(resource, request_body):
 ### @export "request_extension"
 
 @requires_roles('Portal Administrator', 'Agency Administrator',
-                'Agency FOIL Personnel')
+                'Agency FOIL Officer')
 def request_extension(
         request_id,
         extension_reasons,
@@ -1109,7 +1109,7 @@ def get_responses_chronologically(req):
 
             if note.privacy == 2 and (current_user.is_anonymous
                                       or current_user.role not in ['Portal Administrator'
-                    , 'Agency Administrator', 'Agency FOIL Personnel',
+                    , 'Agency Administrator', 'Agency FOIL Officer',
                                                                    'Agency Helpers']):
                 pass
             else:
@@ -1290,11 +1290,11 @@ def change_privacy_setting(request_id, privacy, field):
 def change_record_privacy(record_id, privacy):
     record = get_obj("Record", record_id)
     app.logger.info('Syncing privacy changes to %s' % app.config['PUBLIC_SERVER_HOSTNAME'])
-    if privacy == 'False':
+    if record.filename and privacy == 'False':
         app.logger.info("Making %s public" % record.filename)
         subprocess.call(["mv", app.config['UPLOAD_PRIVATE_LOCAL_FOLDER'] + "/" + record.filename, app.config['UPLOAD_PUBLIC_LOCAL_FOLDER'] + "/"])
         subprocess.call(["rsync", "-avzh", "ssh", app.config['UPLOAD_PUBLIC_LOCAL_FOLDER'] + "/" + record.filename, app.config['PUBLIC_SERVER_USER'] + '@' + app.config['PUBLIC_SERVER_HOSTNAME'] + ':' + app.config['UPLOAD_PUBLIC_REMOTE_FOLDER'] + "/"])
-    elif privacy == 'True':
+    elif record.filename and privacy == 'True':
         app.logger.info("Making %s private" % record.filename)
         subprocess.call(["mv", app.config['UPLOAD_PUBLIC_LOCAL_FOLDER'] + "/" + record.filename, app.config['UPLOAD_PRIVATE_LOCAL_FOLDER'] + "/"])
         subprocess.call(["rsync", "-avzh", "--delete", "ssh", app.config['UPLOAD_PUBLIC_LOCAL_FOLDER'] + "/" + record.filename, app.config['PUBLIC_SERVER_USER'] + '@' + app.config['PUBLIC_SERVER_HOSTNAME'] + ':' + app.config['UPLOAD_PUBLIC_REMOTE_FOLDER'] + "/"])
